# openapi.json
{"openapi":"3.1.0","info":{"title":"Base API","version":"0.1.0"},"paths":{"/api/users/":{"get":{"tags":["users"],"summary":"Get Users","description":"List users. Requires authentication.","operationId":"get_users_api_users__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/User"},"type":"array","title":"Response Get Users Api Users  Get"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["users"],"summary":"Register User","description":"Create a user unless the email is already taken.","operationId":"register_user_api_users__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreatePayload"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["users"],"summary":"Delete Existing User","operationId":"delete_existing_user_api_users__delete","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailPayload"}}},"required":true},"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/users/adminize":{"post":{"tags":["users"],"summary":"Promote User","operationId":"promote_user_api_users_adminize_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailPayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/users/deadminize":{"post":{"tags":["users"],"summary":"Demote User","operationId":"demote_user_api_users_deadminize_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailPayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/":{"get":{"summary":"Read Root","description":"Return a placeholder response for a starting project.","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Read Root  Get"}}}}}}},"/health":{"get":{"summary":"Health","description":"Simple authenticated endpoint for verifying auth.","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Health Get"}}}}},"security":[{"HTTPBearer":[]}]}}},"components":{"schemas":{"EmailPayload":{"properties":{"email":{"type":"string","title":"Email"}},"type":"object","required":["email"],"title":"EmailPayload"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"User":{"properties":{"email":{"type":"string","title":"Email"},"firstname":{"type":"string","title":"Firstname"},"lastname":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Lastname"},"id":{"type":"integer","title":"Id"},"is_admin":{"type":"boolean","title":"Is Admin","default":false},"disabled":{"type":"boolean","title":"Disabled","default":false},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["email","firstname","id","created_at","updated_at"],"title":"User","description":"Representation of a user stored in the in-memory registry."},"UserCreatePayload":{"properties":{"email":{"type":"string","title":"Email"},"firstname":{"type":"string","title":"Firstname"},"lastname":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Lastname"}},"type":"object","required":["email","firstname"],"title":"UserCreatePayload"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}